requirements.txt
pandas
numpy
requests
pyyaml
matplotlib
plotly
yfinance
jinja2
beautifulsoup4
lxml
pytest
